Refactor the client overview into a new component
