+++ grr/server/grr_response_server/gui/ui/components/client/client_test.ts
@@ -1,6 +1,5 @@
 import {Location} from '@angular/common';
 import {async, discardPeriodicTasks, fakeAsync, TestBed, tick} from '@angular/core/testing';
-import {MatChip, MatChipList} from '@angular/material/chips';
 import {MatDrawer} from '@angular/material/sidenav';
 import {By} from '@angular/platform-browser';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
@@ -73,187 +72,6 @@ describe('Client Component', () => {
     expect(searchClientsSpy).toHaveBeenCalledWith('C.1234');
   });
 
-  it('displays client details on client change', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const subject = new Subject<Client>();
-    Object.defineProperty(facade, 'selectedClient$', {get: () => subject});
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    subject.next(newClient({
-      clientId: 'C.1234',
-      knowledgeBase: {
-        fqdn: 'foo.unknown',
-      },
-      lastSeenAt: new Date(1571789996678),
-    }));
-    fixture.detectChanges();
-
-    const text = fixture.debugElement.nativeElement.textContent;
-    expect(text).toContain('C.1234');
-    expect(text).toContain('foo.unknown');
-  });
-
-  it('allows removal of client labels', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const subject = new Subject<Client>();
-    Object.defineProperty(facade, 'selectedClient$', {get: () => subject});
-    spyOn(facade, 'removeClientLabel');
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    subject.next(newClient({
-      clientId: 'C.1234',
-      labels: [{name: 'testlabel', owner: ''}],
-    }));
-    fixture.detectChanges();
-
-    const labelsChipList = fixture.debugElement.query(By.directive(MatChipList))
-                               .componentInstance.chips.toArray() as MatChip[];
-    labelsChipList[0].remove();
-    expect(facade.removeClientLabel).toHaveBeenCalledWith('testlabel');
-  });
-
-  it('shows a snackbar when a client label is removed', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const clientSubject = new Subject<Client>();
-    const removedLabelsSubject = new Subject<string>();
-    Object.defineProperty(
-        facade, 'selectedClient$', {get: () => clientSubject});
-    Object.defineProperty(
-        facade, 'removedClientLabels$', {get: () => removedLabelsSubject});
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    clientSubject.next(newClient({
-      clientId: 'C.1234',
-      labels: [{name: 'testlabel', owner: ''}],
-    }));
-    fixture.detectChanges();
-
-    const labelsChipList = fixture.debugElement.query(By.directive(MatChipList))
-                               .componentInstance.chips.toArray() as MatChip[];
-    labelsChipList[0].remove();
-    removedLabelsSubject.next('testlabel');
-    fixture.detectChanges();
-
-    const snackbarDiv = document.querySelector('snack-bar-container');
-    expect(snackbarDiv).toBeTruthy();
-    expect(snackbarDiv!.textContent).toContain('Label "testlabel" removed');
-    snackbarDiv!.remove();
-  });
-
-  it('doesn\'t show a snackbar when a client label request errors', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const clientSubject = new Subject<Client>();
-    const removedLabelsSubject = new Subject<string>();
-    Object.defineProperty(
-        facade, 'selectedClient$', {get: () => clientSubject});
-    Object.defineProperty(
-        facade, 'removedClientLabels$', {get: () => removedLabelsSubject});
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    clientSubject.next(newClient({
-      clientId: 'C.1234',
-      labels: [{name: 'testlabel', owner: ''}],
-    }));
-    fixture.detectChanges();
-
-    const labelsChipList = fixture.debugElement.query(By.directive(MatChipList))
-                               .componentInstance.chips.toArray() as MatChip[];
-    labelsChipList[0].remove();
-    removedLabelsSubject.error(new Error('Some error'));
-    fixture.detectChanges();
-
-    const snackbarDiv = document.querySelector('snack-bar-container');
-    expect(snackbarDiv).not.toBeTruthy();
-  });
-
-  it('snackbar action undoes a removal of client label', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const clientSubject = new Subject<Client>();
-    const removedLabelsSubject = new Subject<string>();
-    Object.defineProperty(
-        facade, 'selectedClient$', {get: () => clientSubject});
-    Object.defineProperty(
-        facade, 'removedClientLabels$', {get: () => removedLabelsSubject});
-    spyOn(facade, 'addClientLabel');
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    clientSubject.next(newClient({
-      clientId: 'C.1234',
-      labels: [{name: 'testlabel', owner: ''}],
-    }));
-    fixture.detectChanges();
-
-    const labelsChipList = fixture.debugElement.query(By.directive(MatChipList))
-                               .componentInstance.chips.toArray() as MatChip[];
-    labelsChipList[0].remove();
-    removedLabelsSubject.next('testlabel');
-    fixture.detectChanges();
-
-    expect(facade.addClientLabel).not.toHaveBeenCalled();
-
-    const snackbarDivButton =
-        document.querySelector('div.mat-simple-snackbar-action button');
-    snackbarDivButton!.dispatchEvent(new MouseEvent('click'));
-    fixture.detectChanges();
-    fixture.whenRenderingDone().then(() => {
-      expect(facade.addClientLabel).toHaveBeenCalledWith('testlabel');
-    });
-  });
-
-  it('allows viewing of client details', () => {
-    // Prevent warnings from 404-ing API requests.
-    spyOn(facade, 'selectClient');
-
-    const subject = new Subject<Client>();
-    Object.defineProperty(facade, 'selectedClient$', {get: () => subject});
-    spyOn(facade, 'removeClientLabel');
-
-    const fixture = TestBed.createComponent(ClientComponent);
-    fixture.detectChanges();  // Ensure ngOnInit hook completes.
-
-    paramsSubject.next(new Map(Object.entries({id: 'C.1234'})));
-    subject.next(newClient({
-      clientId: 'C.1234',
-      labels: [{name: 'testlabel', owner: ''}],
-    }));
-    fixture.detectChanges();
-    const detailsDrawer =
-        fixture.debugElement.query(By.directive(MatDrawer)).componentInstance;
-    const drawerOpenSpy = spyOn(detailsDrawer, 'open');
-
-    const detailsButton =
-        fixture.debugElement.query(By.css('.goto-details')).nativeElement;
-    detailsButton.dispatchEvent(new MouseEvent('click'));
-
-    expect(drawerOpenSpy).toHaveBeenCalled();
-    detailsDrawer.close();
-  });
-
   it('correctly updates URL for client details drawer usage', fakeAsync(() => {
        // Prevent warnings from 404-ing API requests.
        spyOn(facade, 'selectClient');
@@ -306,7 +124,7 @@ describe('Client Component', () => {
 
        // The following expectation is met when testing manually, but not on
        // automated testing, because the drawer's closedStart observable is not
-       // firing
+       // emitting
        // expect(location.path()).toEqual('/v2/clients/C.1234');
        discardPeriodicTasks();
      }));
