+++ grr/server/grr_response_server/gui/ui/components/client/module.ts
@@ -1,22 +1,14 @@
 import {NgModule} from '@angular/core';
-import {MatButtonModule} from '@angular/material/button';
-import {MatChipsModule} from '@angular/material/chips';
-import {MatIconModule} from '@angular/material/icon';
-import {MatSnackBarModule} from '@angular/material/snack-bar';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {RouterModule} from '@angular/router';
 import {FlowFormModule} from '@app/components/flow_form/module';
 import {FlowListModule} from '@app/components/flow_list/module';
 
 import {ApprovalModule} from '../approval/module';
-import {ClientAddLabelDialogModule} from '../client_add_label_dialog/module';
-import {OnlineChipModule} from '../online_chip/module';
-import {TimestampModule} from '../timestamp/module';
+import {ClientOverviewModule} from '../client_overview/module';
 
 import {Client} from './client';
 import {ClientRoutingModule} from './routing';
-import {ClientDetailsModule} from '../client_details/module';
-import {MatSidenavModule} from '@angular/material/sidenav';
 
 /**
  * Module for the client details component.
@@ -25,20 +17,11 @@ import {MatSidenavModule} from '@angular/material/sidenav';
   imports: [
     BrowserAnimationsModule,
     RouterModule,
-    MatChipsModule,
     ClientRoutingModule,
     ApprovalModule,
     FlowFormModule,
     FlowListModule,
-    TimestampModule,
-    OnlineChipModule,
-    MatButtonModule,
-    MatIconModule,
-    MatButtonModule,
-    ClientAddLabelDialogModule,
-    MatSnackBarModule,
-    ClientDetailsModule,
-    MatSidenavModule,
+    ClientOverviewModule,
   ],
   declarations: [
     Client,
